Fix, used int instead of long
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/plugins/Freetalk@26923 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/plugins/Freetalk/Board.java
@@ -498,7 +498,7 @@ public final class Board implements Comparable<Board> {
     }
 
     @SuppressWarnings("unchecked")
-    public synchronized List<MessageReference> getMessagesByMinimumDate(int minimumDate, final boolean sortByMessageDateAscending) {
+    public synchronized List<MessageReference> getMessagesByMinimumDate(long minimumDate, final boolean sortByMessageDateAscending) {
         final Query q = db.query();
         q.constrain(BoardMessageLink.class);
         q.descend("mBoard").constrain(this);
